iteration 3

+++ server/routes/rooms.route.js
@@ -1,7 +1,8 @@
 const router = require("express").Router();
 const RoomModel = require("../models/Room.model");
+const ReviewModel = require("../models/Review.model");
 
-router.post("/room", async (req, res) => {
+router.post("/room", isAuthenticated, async (req, res) => {
   try {
     const data = req.body;
     const result = await RoomModel.create(data);
@@ -13,22 +14,22 @@ router.post("/room", async (req, res) => {
   }
 });
 
-router.get("/room", async (req, res) => {
+router.get("/room", isAuthenticated, async (req, res) => {
   try {
     const rooms = await RoomModel.find();
 
     return res.status(200).json(rooms);
   } catch (err) {
     console.error(err);
-    return res.status(500).json({ msg: "Falha ao buscar todos os quartos" });
+    return res.status(500).json({ msg: "Falha ao buscar os quartos" });
   }
 });
 
-router.get("/room/:_id", async (req, res) => {
+router.get("/room/:_id", isAuthenticated, async (req, res) => {
   try {
     const { _id } = req.params;
 
-    const room = await RoomModel.findOne({ _id });
+    const room = await RoomModel.findOne({ _id }).populate("reviews");
 
     if (!room) {
       return res.status(404).json({ msg: "Quarto não encontrado" });
@@ -41,13 +42,13 @@ router.get("/room/:_id", async (req, res) => {
   }
 });
 
-router.patch("/room/:_id", async (req, res) => {
+router.patch("/room/:_id", isAuthenticated, async (req, res) => {
   try {
     const { _id } = req.params;
     const data = req.body;
 
     const result = await RoomModel.findOneAndUpdate(
-      { _id },
+      { _id, ownerId: req.user._id },
       { $set: data },
       { new: true, runValidators: true }
     );
@@ -55,23 +56,26 @@ router.patch("/room/:_id", async (req, res) => {
       return res.status(404).json({ msg: "Quarto não encontrado" });
     }
 
-    return res.status(202).json(room);
+    return res.status(202).json(result);
   } catch (err) {
     console.error(err);
     return res.status(500).json({ msg: "Falha ao editar o quarto" });
   }
 });
 
-router.delete("/room/:_id", async (req, res) => {
+router.delete("/room/:_id", isAuthenticated, async (req, res) => {
   try {
     const { _id } = req.params;
 
-    const result = await RoomModel.deleteOne({ _id });
+    const result = await RoomModel.deleteOne({ _id, ownerId: req.user._id });
+
+    const reviewResult = await ReviewModel.deleteMany({ roomId: _id });
 
     if (result.deletedCount < 1) {
       return res.status(404).json({ msg: "Quarto não encontrado" });
     }
 
+    console.log(reviewResult);
     return res.status(200).json({});
   } catch (err) {
     console.error(err);
